LLM-projects/sample_projects/RAG-gatway-service/requirements.txt
fastapi==0.109.2
httpx==0.26.0
pydantic==2.6.1
pyyaml==6.0.1
uvicorn==0.27.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dotenv==1.0.1
openai==1.12.0 